# openapi/kcp-capsules-and-export.yaml
openapi: 3.0.3
info:
  title: KYBa Gitea Capsule and Export API
  version: 0.2.0
  description: >-
    Draft API surface for the Gitea fork overlay. Repository interfaces,
    capsules and multirepo archive export are first-class platform resources.
servers:
  - url: /api/v1
paths:
  /kcp/capsules:
    get:
      summary: List exported capsules visible to the caller.
      parameters:
        - name: owner_repo
          in: query
          schema: { type: string }
        - name: kind
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Capsule list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  capsules:
                    type: array
                    items:
                      $ref: '#/components/schemas/CapsuleManifest'
    post:
      summary: Publish or update a repository-owned capsule export.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapsuleManifest'
      responses:
        '202':
          description: Capsule accepted for validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapsuleStatus'
...
  /kcp/export-plan:
    post:
      summary: Validate a monorepo export manifest and return file ownership plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportPlanRequest'
      responses:
        '200':
          description: Export plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPlan'
        '409':
          description: Ambiguous or unassigned ownership blocks the export.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPlan'
...
components:
...
  schemas:
    CapsuleManifest:
      type: object
      required: [id, kind, owner_repo, version, visibility, exports]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [product, screen, api, event, ci, context, repository-interface]
        owner_repo: { type: string }
        version: { type: string }
        visibility:
          type: string
          enum: [private, imported-repos-only, public]
        exports:
          type: array
          items: { type: string }
        consumers:
          type: array
          items: { type: string }
        impact_policy:
          type: string
          enum: [create-generated-draft-pr, create-maintenance-task, block-until-owner-approval]
        freshness:
          type: object
          properties:
            max_age_days: { type: integer, minimum: 0 }
    CapsuleStatus:
      type: object
      properties:
        capsule_id: { type: string }
        status: { type: string }
...
    ExportPlanRequest:
      type: object
      required: [manifest, files]
      properties:
        manifest: { type: object }
        files:
          type: array
          items: { type: string }
    ExportPlan:
      type: object
      properties:
        ready: { type: boolean }
        import_order:
          type: array
          items: { type: string }
        ownership_digest: { type: string }
        assignments:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              target: { type: string }
        ambiguous_files:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        unassigned_files:
          type: array
          items: { type: string }
...